+++ tests/Builder/CreateStatementTest.php
@@ -28,6 +28,7 @@ class CreateStatementTest extends TestCase
 
     public function testBuilderDatabase()
     {
+        // CREATE DATABASE ...
         $parser = new Parser(
             'CREATE DATABASE `mydb` ' .
             'DEFAULT CHARACTER SET = utf8 DEFAULT COLLATE = utf8_general_ci'
@@ -36,6 +37,20 @@ class CreateStatementTest extends TestCase
 
         $this->assertEquals(
             'CREATE DATABASE `mydb` ' .
+            'DEFAULT CHARACTER SET=utf8 DEFAULT COLLATE=utf8_general_ci',
+            $stmt->build()
+        );
+
+
+        // CREATE SCHEMA ...
+        $parser = new Parser(
+            'CREATE SCHEMA `mydb` ' .
+            'DEFAULT CHARACTER SET = utf8 DEFAULT COLLATE = utf8_general_ci'
+        );
+        $stmt = $parser->statements[0];
+
+        $this->assertEquals(
+            'CREATE SCHEMA `mydb` ' .
             'DEFAULT CHARACTER SET=utf8 DEFAULT COLLATE=utf8_general_ci',
             $stmt->build()
         );